Admin | Create product

+++ database/migrations/2018_06_30_170035_create_products_table.php
@@ -17,7 +17,7 @@ class CreateProductsTable extends Migration
             $table->increments('id');
             $table->string('name');
             $table->text('description');
-            $table->integer('price');
+            $table->decimal('price', 7, 2);
             $table->string('image');
             $table->timestamps();
         });
